Remove `ColorSpace.parseAsync` and simplify the ColorSpace "API-surface"
This patch reduces the number of `ColorSpace` static-methods, and in particular the `parseAsync` method is removed and it's now instead possible to have `parse` optionally return a Promise.
This thus removes the need to manually check if a `ColorSpace`-instance is cached, note the changes in the `src/core/evaluator.js` file.